+++ lib/matestack/ui/core/app.rb
@@ -5,6 +5,7 @@ module Matestack
         
         def initialize(options = {})
           @controller = Context.controller
+          Context.app = self
           super(nil, nil, options)
         end
 
@@ -16,6 +17,17 @@ module Matestack
         end
 
         def loading_state_element
+        end
+
+
+        # layout class method to specify if a rails layout should be used
+        def self.inherited(subclass)
+          subclass.layout(@layout)
+          super
+        end
+          
+        def self.layout(layout = nil)
+          @layout = layout ? layout : @layout
         end
         
       end