+++ lib/matestack/ui/core/context.rb
@@ -3,6 +3,7 @@ module Matestack
     module Core
       class Context < ActiveSupport::CurrentAttributes
 
+        attribute :app
         attribute :parent
         attribute :isolated_parent
         attribute :params