+++ lib/matestack/ui/core/helper.rb
@@ -19,25 +19,26 @@ module Matestack
         end
         
         def render(*args)
+          setup_context
           if args.first.is_a?(Class) && args.first.ancestors.include?(Base)
-            setup_context
             raise 'expected a hash as second argument' unless args.second.is_a?(Hash) || args.second.nil?
             options = args.second || {}
-            app = options.delete(:app) || self.class.matestack_app
+            app = options.delete(:matestack_app) || self.class.matestack_app
+            layout = app ? app.layout : false
             page = args.first
             if app && params[:only_page].nil? && params[:component_key].nil? && params[:component_class].nil?
-              render_app app, page, options
+              render_app app, page, options, layout
             else
               if params[:component_key] && params[:component_class].nil?
-                render_component app, page, params[:component_key], options
+                render_component app, page, params[:component_key], options, layout
               elsif params[:component_class]
                 if params[:component_key]
-                  render_component nil, params[:component_class].constantize, params[:component_key], JSON.parse(params[:public_options] || '{}')
+                  render_component nil, params[:component_class].constantize, params[:component_key], JSON.parse(params[:public_options] || '{}'), layout
                 else 
                   render html: params[:component_class].constantize.(public_options: JSON.parse(params[:public_options] || '{}'))
                 end
               else
-                render_page page, options
+                render_page page, options, layout
               end
             end
           else
@@ -45,22 +46,22 @@ module Matestack
           end
         end
         
-        def render_app(app, page, options)
-          render html: app.new(options) { page.new(options) }.render_content.html_safe, layout: false
+        def render_app(app, page, options, layout)
+          render html: app.new(options) { page.new(options) }.render_content.html_safe, layout: layout
         end
         
-        def render_page(page, options)
-          render html: page.new(options).render_content.html_safe, layout: false
+        def render_page(page, options, layout)
+          render html: page.new(options).render_content.html_safe, layout: layout
         end
         
-        def render_component(app, page, component_key, options)
+        def render_component(app, page, component_key, options, layout)
           app ? app.new(options) { page.new(options) } : page.new(options) # create page structure in order to later access registered async components
-          render html: Matestack::Ui::Core::Context.async_components[component_key].render_content.html_safe, layout: false
+          render html: Matestack::Ui::Core::Context.async_components[component_key].render_content.html_safe, layout: layout
         end
         
         def setup_context
           Matestack::Ui::Core::Context.params = self.params
-          Matestack::Ui::Core::Context.controller = self
+          Matestack::Ui::Core::Context.controller = (self.class <= ActionController::Base) ? self : @_controller
         end
         
       end