+++ spec/dummy/app/helpers/application_helper.rb
@@ -1,3 +1,4 @@
 module ApplicationHelper
-  # include Matestack::Ui::Core::ApplicationHelper
+  include Matestack::Ui::Core::Helper
+  include Matestack::Ui::Core::TagHelper
 end