+++ spec/dummy/app/javascript/packs/application.js
@@ -15,6 +15,7 @@ import Vuex from 'vuex'
 
 import MatestackUiCore from 'matestack-ui-core'
 window.MatestackUiCore = MatestackUiCore // making MatestackUiCore globally available for test compatability
+MatestackUiCore.Vue = Vue // test compatability
 let matestackUiApp = undefined
 
 document.addEventListener('DOMContentLoaded', () => {