+++ spec/dummy/app/matestack/components/legacy_views/pages/async.rb
@@ -1,13 +1,13 @@
-class Components::LegacyViews::Pages::Async < Matestack::Ui::StaticComponent
+class Components::LegacyViews::Pages::Async < Matestack::Ui::Component
 
   def response
     div id: 'foobar' do
-      paragraph text: 'Im a custom component'
+      paragraph 'Im a custom component'
       async rerender_on: 'update_time', id: 'async-legacy-integratable' do 
-        paragraph text: I18n.l(DateTime.now)
+        paragraph I18n.l(DateTime.now)
       end
       onclick emit: 'update_time' do
-        button text: 'Click me!'
+        button 'Click me!'
       end
     end
   end