+++ spec/dummy/app/matestack/components/legacy_views/pages/onclick.rb
@@ -1,8 +1,8 @@
-class Components::LegacyViews::Pages::Onclick < Matestack::Ui::StaticComponent
+class Components::LegacyViews::Pages::Onclick < Matestack::Ui::Component
 
   def response
     onclick emit: 'update_time' do
-      button text: 'Click me!'
+      button 'Click me!'
     end
     toggle show_on: 'update_time', hide_after: 1000, id: 'async-onclick' do
       plain 'clicked'